pyproject.toml
[project]
name = "jumpingjay"
version = "0.1.0"
description = "Calculate duration since a given time"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["python-dateutil>=2.8.0"]

[project.scripts]
durationsince = "jumpingjay.main:durationsince"
durationtill = "jumpingjay.main:durationtill"
jumpingjay = "jumpingjay.main:list_commands"

[tool.ruff]
preview = true

[tool.ruff.lint]
select = [
  "A",
  "AIR",
  "ARG",
  "ASYNC",
  "ASYNC1",
  "B",
  "C",
  "C4",
  "C90",
  "DJ",
  "DTZ",
  "E",
  "EM",
  "ERA",
  "EXE",
  "F",
  "FA",
  "FBT",
  "FIX",
  "FLY",
  "G",
  "I",
  "ICN",
  "INP",
  "INT",
  "ISC",
  "LOG",
  "N",
  "NPY",
  "PD",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "Q",
  "R",
  "RET",
  "RSE",
  "RUF",
  "S",
  "SIM",
  "SLF",
  "SLOT",
  "T10",
  "T20",
  "TD",
  "TID",
  "TRY",
  "UP",
  "W",
  "YTT",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
